feat(liquidity): Init
