+++ src/views/BmpHome/index.config.ts
@@ -1,0 +1,3 @@
+export default {
+  navigationStyle: 'custom',
+}
