+++ src/views/Swap/bmp/index.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { useTracker } from 'contexts/AnalyticsContext'
 import { HitBuilders } from 'utils/ga'
-import mpService from '@binance/mp-service'
 import styled, { useTheme } from 'styled-components'
 import { CurrencyAmount, JSBI, Token, Trade } from '@pancakeswap/sdk'
 import {
@@ -12,20 +11,12 @@ import {
   useModal,
   Flex,
   IconButton,
-  BottomDrawer,
   useMatchBreakpoints,
   ArrowUpDownIcon,
-  Modal,
-  WalletIcon,
-  SwapFillIcon,
 } from '@pancakeswap/uikit'
 import { useIsTransactionUnsupported } from 'hooks/Trades'
-import UnsupportedCurrencyFooter from 'components/UnsupportedCurrencyFooter'
-import Footer from 'components/Menu/Footer'
-import { useRouter } from 'next/router'
 import { useTranslation } from 'contexts/Localization'
 import SwapWarningTokens from 'config/constants/swapWarningTokens'
-import TransactionsModal from 'components/App/Transactions/TransactionsModal'
 import { useDisclaimer } from 'hooks/useDisclaimer'
 
 import AddressInputPanel from '../components/AddressInputPanel'
@@ -34,7 +25,6 @@ import Column, { AutoColumn } from '../../../components/Layout/Column'
 import ConfirmSwapModal from '../components/ConfirmSwapModal'
 import CurrencyInputPanel from '../../../components/CurrencyInputPanel'
 import { AutoRow, RowBetween } from '../../../components/Layout/Row'
-import AdvancedSwapDetailsDropdown from '../components/AdvancedSwapDetailsDropdown'
 import confirmPriceImpactWithoutFee from '../components/confirmPriceImpactWithoutFee'
 import { ArrowWrapper, SwapCallbackError, Wrapper } from '../components/styleds'
 import TradePrice from '../components/TradePrice'
@@ -66,12 +56,10 @@ import { maxAmountSpend } from '../../../utils/maxAmountSpend'
 import { computeTradePriceBreakdown, warningSeverity } from '../../../utils/prices'
 import CircleLoader from '../../../components/Loader/CircleLoader'
 import ErrorBoundary from '../../../components/ErrorBoundary'
-import Page from '../../Page'
 import SwapWarningModal from '../components/SwapWarningModal'
 import PriceChartContainer from '../components/Chart/PriceChartContainer'
 import { StyledInputCurrencyWrapper, StyledSwapContainer } from '../styles'
 import CurrencyInputHeader from '../components/CurrencyInputHeader'
-import Providers from '../../../PageProvider.bmp'
 
 const Label = styled(Text)`
   font-size: 12px;
@@ -96,7 +84,6 @@ const SwitchIconButton = styled(IconButton)`
   }
 `
 function Swap() {
-  const router = useRouter()
   const loadedUrlParams = useDefaultsFromURLSearch()
   const { t } = useTranslation()
   const { isMobile } = useMatchBreakpoints()
@@ -380,176 +367,146 @@ function Swap() {
   )
 
   return (
-    <Page>
-      <Flex flexDirection="column">
-        <StyledSwapContainer $isChartExpanded={isChartExpanded}>
-          <StyledInputCurrencyWrapper mt={isChartExpanded ? '24px' : '0'} mr={isChartExpanded ? '0' : '0'}>
-            <AppBody>
-              <CurrencyInputHeader
-                title={t('Swap')}
-                subtitle={t('Trade tokens in an instant')}
-                setIsChartDisplayed={setIsChartDisplayed}
-                isChartDisplayed={isChartDisplayed}
-              />
-              <Wrapper id="swap-page" style={{ minHeight: '412px' }}>
-                <AutoColumn gap="sm">
-                  <CurrencyInputPanel
-                    label={independentField === Field.OUTPUT && !showWrap && trade ? t('From (estimated)') : t('From')}
-                    value={formattedAmounts[Field.INPUT]}
-                    showMaxButton={!atMaxAmountInput}
-                    currency={currencies[Field.INPUT]}
-                    onUserInput={handleTypeInput}
-                    onMax={handleMaxInput}
-                    onCurrencySelect={handleInputSelect}
-                    otherCurrency={currencies[Field.OUTPUT]}
-                    id="swap-currency-input"
-                  />
+    <Flex flexDirection="column">
+      <StyledSwapContainer $isChartExpanded={isChartExpanded}>
+        <StyledInputCurrencyWrapper mt={isChartExpanded ? '24px' : '0'} mr={isChartExpanded ? '0' : '0'}>
+          <AppBody>
+            <CurrencyInputHeader
+              title={t('Swap')}
+              subtitle={t('Trade tokens in an instant')}
+              setIsChartDisplayed={setIsChartDisplayed}
+              isChartDisplayed={isChartDisplayed}
+            />
+            <Wrapper id="swap-page" style={{ minHeight: '412px' }}>
+              <AutoColumn gap="sm">
+                <CurrencyInputPanel
+                  label={independentField === Field.OUTPUT && !showWrap && trade ? t('From (estimated)') : t('From')}
+                  value={formattedAmounts[Field.INPUT]}
+                  showMaxButton={!atMaxAmountInput}
+                  currency={currencies[Field.INPUT]}
+                  onUserInput={handleTypeInput}
+                  onMax={handleMaxInput}
+                  onCurrencySelect={handleInputSelect}
+                  otherCurrency={currencies[Field.OUTPUT]}
+                  id="swap-currency-input"
+                />
 
-                  <AutoColumn justify="space-between">
-                    <AutoRow justify={isExpertMode ? 'space-between' : 'center'} style={{ padding: '0 12px' }}>
-                      <SwitchIconButton
-                        variant="light"
-                        scale="sm"
-                        onClick={() => {
-                          setApprovalSubmitted(false) // reset 2 step UI for approvals
-                          onSwitchTokens()
-                        }}
-                      >
-                        <ArrowDownIcon
-                          className="icon-down"
-                          color={currencies[Field.INPUT] && currencies[Field.OUTPUT] ? 'primary' : 'text'}
-                        />
-                        <ArrowUpDownIcon className="icon-up-down" color="#ffffff" />
-                      </SwitchIconButton>
-                      {recipient === null && !showWrap && isExpertMode ? (
-                        <Button variant="text" id="add-recipient-button" onClick={() => onChangeRecipient('')}>
-                          {t('+ Add a send (optional)')}
-                        </Button>
-                      ) : null}
-                    </AutoRow>
-                  </AutoColumn>
-                  <CurrencyInputPanel
-                    value={formattedAmounts[Field.OUTPUT]}
-                    onUserInput={handleTypeOutput}
-                    label={independentField === Field.INPUT && !showWrap && trade ? t('To (estimated)') : t('To')}
-                    showMaxButton={false}
-                    currency={currencies[Field.OUTPUT]}
-                    onCurrencySelect={handleOutputSelect}
-                    otherCurrency={currencies[Field.INPUT]}
-                    id="swap-currency-output"
-                  />
-
-                  {isExpertMode && recipient !== null && !showWrap ? (
-                    <>
-                      <AutoRow justify="space-between" style={{ padding: '0 1rem' }}>
-                        <ArrowWrapper clickable={false}>
-                          <ArrowDownIcon width="16px" />
-                        </ArrowWrapper>
-                        <Button variant="text" id="remove-recipient-button" onClick={() => onChangeRecipient(null)}>
-                          {t('- Remove send')}
-                        </Button>
-                      </AutoRow>
-                      <AddressInputPanel id="recipient" value={recipient} onChange={onChangeRecipient} />
-                    </>
-                  ) : null}
-
-                  {showWrap ? null : (
-                    <AutoColumn gap="8px" style={{ padding: '0 16px' }}>
-                      <RowBetween align="center">
-                        {Boolean(trade) && (
-                          <>
-                            <Label>{t('Price')}</Label>
-                            <TradePrice
-                              price={trade?.executionPrice}
-                              showInverted={showInverted}
-                              setShowInverted={setShowInverted}
-                            />
-                          </>
-                        )}
-                      </RowBetween>
-                      <RowBetween align="center">
-                        <Label>{t('Slippage Tolerance')}</Label>
-                        <Text bold color="primary">
-                          {allowedSlippage / 100}%
-                        </Text>
-                      </RowBetween>
-                    </AutoColumn>
-                  )}
+                <AutoColumn justify="space-between">
+                  <AutoRow justify={isExpertMode ? 'space-between' : 'center'} style={{ padding: '0 12px' }}>
+                    <SwitchIconButton
+                      variant="light"
+                      scale="sm"
+                      onClick={() => {
+                        setApprovalSubmitted(false) // reset 2 step UI for approvals
+                        onSwitchTokens()
+                      }}
+                    >
+                      <ArrowDownIcon
+                        className="icon-down"
+                        color={currencies[Field.INPUT] && currencies[Field.OUTPUT] ? 'primary' : 'text'}
+                      />
+                      <ArrowUpDownIcon className="icon-up-down" color="#ffffff" />
+                    </SwitchIconButton>
+                    {recipient === null && !showWrap && isExpertMode ? (
+                      <Button variant="text" id="add-recipient-button" onClick={() => onChangeRecipient('')}>
+                        {t('+ Add a send (optional)')}
+                      </Button>
+                    ) : null}
+                  </AutoRow>
                 </AutoColumn>
-                <Box mt="0.25rem">
-                  {swapIsUnsupported ? (
-                    <Button width="100%" disabled>
-                      {t('Unsupported Asset')}
-                    </Button>
-                  ) : !account ? (
-                    <ConnectWalletButton width="100%" />
-                  ) : showWrap ? (
-                    <Button width="100%" disabled={Boolean(wrapInputError)} onClick={onWrap}>
-                      {wrapInputError ??
-                        (wrapType === WrapType.WRAP ? 'Wrap' : wrapType === WrapType.UNWRAP ? 'Unwrap' : null)}
-                    </Button>
-                  ) : noRoute && userHasSpecifiedInputOutput ? (
-                    <GreyCard style={{ textAlign: 'center', padding: '0.75rem' }}>
-                      <Text color="textSubtle">{t('Insufficient liquidity for this trade.')}</Text>
-                      {singleHopOnly && <Text color="textSubtle">{t('Try enabling multi-hop trades.')}</Text>}
-                    </GreyCard>
-                  ) : showApproveFlow ? (
-                    <RowBetween>
-                      <Button
-                        variant={approval === ApprovalState.APPROVED ? 'success' : 'primary'}
-                        onClick={approveCallback}
-                        disabled={approval !== ApprovalState.NOT_APPROVED || approvalSubmitted}
-                        width="48%"
-                      >
-                        {approval === ApprovalState.PENDING ? (
-                          <AutoRow gap="6px" justify="center">
-                            {t('Enabling')} <CircleLoader stroke="white" />
-                          </AutoRow>
-                        ) : approvalSubmitted && approval === ApprovalState.APPROVED ? (
-                          t('Enabled')
-                        ) : (
-                          t('Enable %asset%', { asset: currencies[Field.INPUT]?.symbol ?? '' })
-                        )}
-                      </Button>
-                      <Button
-                        variant={isValid && priceImpactSeverity > 2 ? 'danger' : 'primary'}
-                        onClick={() => {
-                          if (isExpertMode) {
-                            handleSwap()
-                          } else {
-                            tracker.send(
-                              new HitBuilders.EventBuilder().setCategory('swap').setAction('clickSwap').build(),
-                            )
+                <CurrencyInputPanel
+                  value={formattedAmounts[Field.OUTPUT]}
+                  onUserInput={handleTypeOutput}
+                  label={independentField === Field.INPUT && !showWrap && trade ? t('To (estimated)') : t('To')}
+                  showMaxButton={false}
+                  currency={currencies[Field.OUTPUT]}
+                  onCurrencySelect={handleOutputSelect}
+                  otherCurrency={currencies[Field.INPUT]}
+                  id="swap-currency-output"
+                />
 
-                            setSwapState({
-                              tradeToConfirm: trade,
-                              attemptingTxn: false,
-                              swapErrorMessage: undefined,
-                              txHash: undefined,
-                            })
-                            onPresentConfirmModal()
-                          }
-                        }}
-                        width="48%"
-                        id="swap-button"
-                        disabled={
-                          !isValid || approval !== ApprovalState.APPROVED || (priceImpactSeverity > 3 && !isExpertMode)
-                        }
-                      >
-                        {priceImpactSeverity > 3 && !isExpertMode
-                          ? t('Price Impact High')
-                          : priceImpactSeverity > 2
-                          ? t('Swap Anyway')
-                          : t('Swap')}
+                {isExpertMode && recipient !== null && !showWrap ? (
+                  <>
+                    <AutoRow justify="space-between" style={{ padding: '0 1rem' }}>
+                      <ArrowWrapper clickable={false}>
+                        <ArrowDownIcon width="16px" />
+                      </ArrowWrapper>
+                      <Button variant="text" id="remove-recipient-button" onClick={() => onChangeRecipient(null)}>
+                        {t('- Remove send')}
                       </Button>
+                    </AutoRow>
+                    <AddressInputPanel id="recipient" value={recipient} onChange={onChangeRecipient} />
+                  </>
+                ) : null}
+
+                {showWrap ? null : (
+                  <AutoColumn gap="8px" style={{ padding: '0 16px' }}>
+                    <RowBetween align="center">
+                      {Boolean(trade) && (
+                        <>
+                          <Label>{t('Price')}</Label>
+                          <TradePrice
+                            price={trade?.executionPrice}
+                            showInverted={showInverted}
+                            setShowInverted={setShowInverted}
+                          />
+                        </>
+                      )}
                     </RowBetween>
-                  ) : (
+                    <RowBetween align="center">
+                      <Label>{t('Slippage Tolerance')}</Label>
+                      <Text bold color="primary">
+                        {allowedSlippage / 100}%
+                      </Text>
+                    </RowBetween>
+                  </AutoColumn>
+                )}
+              </AutoColumn>
+              <Box mt="0.25rem">
+                {swapIsUnsupported ? (
+                  <Button width="100%" disabled>
+                    {t('Unsupported Asset')}
+                  </Button>
+                ) : !account ? (
+                  <ConnectWalletButton width="100%" />
+                ) : showWrap ? (
+                  <Button width="100%" disabled={Boolean(wrapInputError)} onClick={onWrap}>
+                    {wrapInputError ??
+                      (wrapType === WrapType.WRAP ? 'Wrap' : wrapType === WrapType.UNWRAP ? 'Unwrap' : null)}
+                  </Button>
+                ) : noRoute && userHasSpecifiedInputOutput ? (
+                  <GreyCard style={{ textAlign: 'center', padding: '0.75rem' }}>
+                    <Text color="textSubtle">{t('Insufficient liquidity for this trade.')}</Text>
+                    {singleHopOnly && <Text color="textSubtle">{t('Try enabling multi-hop trades.')}</Text>}
+                  </GreyCard>
+                ) : showApproveFlow ? (
+                  <RowBetween>
                     <Button
-                      variant={isValid && priceImpactSeverity > 2 && !swapCallbackError ? 'danger' : 'primary'}
+                      variant={approval === ApprovalState.APPROVED ? 'success' : 'primary'}
+                      onClick={approveCallback}
+                      disabled={approval !== ApprovalState.NOT_APPROVED || approvalSubmitted}
+                      width="48%"
+                    >
+                      {approval === ApprovalState.PENDING ? (
+                        <AutoRow gap="6px" justify="center">
+                          {t('Enabling')} <CircleLoader stroke="white" />
+                        </AutoRow>
+                      ) : approvalSubmitted && approval === ApprovalState.APPROVED ? (
+                        t('Enabled')
+                      ) : (
+                        t('Enable %asset%', { asset: currencies[Field.INPUT]?.symbol ?? '' })
+                      )}
+                    </Button>
+                    <Button
+                      variant={isValid && priceImpactSeverity > 2 ? 'danger' : 'primary'}
                       onClick={() => {
                         if (isExpertMode) {
                           handleSwap()
                         } else {
+                          tracker.send(
+                            new HitBuilders.EventBuilder().setCategory('swap').setAction('clickSwap').build(),
+                          )
+
                           setSwapState({
                             tradeToConfirm: trade,
                             attemptingTxn: false,
@@ -559,91 +516,66 @@ function Swap() {
                           onPresentConfirmModal()
                         }
                       }}
+                      width="48%"
                       id="swap-button"
-                      width="100%"
-                      disabled={!isValid || (priceImpactSeverity > 3 && !isExpertMode) || !!swapCallbackError}
+                      disabled={
+                        !isValid || approval !== ApprovalState.APPROVED || (priceImpactSeverity > 3 && !isExpertMode)
+                      }
                     >
-                      {swapInputError ||
-                        (priceImpactSeverity > 3 && !isExpertMode
-                          ? t('Price Impact Too High')
-                          : priceImpactSeverity > 2
-                          ? t('Swap Anyway')
-                          : t('Swap'))}
+                      {priceImpactSeverity > 3 && !isExpertMode
+                        ? t('Price Impact High')
+                        : priceImpactSeverity > 2
+                        ? t('Swap Anyway')
+                        : t('Swap')}
                     </Button>
-                  )}
-                  {showApproveFlow && (
-                    <Column style={{ marginTop: '1rem' }}>
-                      <ProgressSteps steps={[approval === ApprovalState.APPROVED]} />
-                    </Column>
-                  )}
-                  {isExpertMode && swapErrorMessage ? <SwapCallbackError error={swapErrorMessage} /> : null}
-                </Box>
-              </Wrapper>
-            </AppBody>
-          </StyledInputCurrencyWrapper>
-        </StyledSwapContainer>
-      </Flex>
-    </Page>
-  )
-}
-
-const FooterWrap = styled(Flex)`
-  position: fixed;
-  bottom: 0px;
-  width: 100%;
-  padding: 10px;
-  background: ${({ theme }) => theme.colors.backgroundAlt};
-  padding-bottom: 24px;
-  height: var(--mp-tabbar-height);
-  z-index: 20;
-  justify-content: center;
-`
-const Item = ({ icon, title, isActive, ...props }) => {
-  const theme = useTheme()
-  return (
-    <Flex flexDirection="column" alignItems="center" width="94px" {...props}>
-      <Box mb="5px">
-        {React.createElement(icon, {
-          color: isActive ? theme.colors.secondary : theme.colors.textSubtle,
-        })}
-      </Box>
-      <Box fontSize="14px" fontWeight={isActive ? '600' : '400'} color={isActive ? 'secondary' : 'textSubtle'}>
-        {title}
-      </Box>
+                  </RowBetween>
+                ) : (
+                  <Button
+                    variant={isValid && priceImpactSeverity > 2 && !swapCallbackError ? 'danger' : 'primary'}
+                    onClick={() => {
+                      if (isExpertMode) {
+                        handleSwap()
+                      } else {
+                        setSwapState({
+                          tradeToConfirm: trade,
+                          attemptingTxn: false,
+                          swapErrorMessage: undefined,
+                          txHash: undefined,
+                        })
+                        onPresentConfirmModal()
+                      }
+                    }}
+                    id="swap-button"
+                    width="100%"
+                    disabled={!isValid || (priceImpactSeverity > 3 && !isExpertMode) || !!swapCallbackError}
+                  >
+                    {swapInputError ||
+                      (priceImpactSeverity > 3 && !isExpertMode
+                        ? t('Price Impact Too High')
+                        : priceImpactSeverity > 2
+                        ? t('Swap Anyway')
+                        : t('Swap'))}
+                  </Button>
+                )}
+                {showApproveFlow && (
+                  <Column style={{ marginTop: '1rem' }}>
+                    <ProgressSteps steps={[approval === ApprovalState.APPROVED]} />
+                  </Column>
+                )}
+                {isExpertMode && swapErrorMessage ? <SwapCallbackError error={swapErrorMessage} /> : null}
+              </Box>
+            </Wrapper>
+          </AppBody>
+        </StyledInputCurrencyWrapper>
+      </StyledSwapContainer>
     </Flex>
   )
 }
-const FooterMenu = () => {
-  const { t } = useTranslation()
-  const list = [
-    { icon: SwapFillIcon, title: t('Exchange'), isActive: true },
-    {
-      icon: WalletIcon,
-      title: t('Wallet'),
-      onClick: () => {
-        mpService.navigateToMiniProgram({
-          appId: 'hhL98uho2A4sGYSHCEdCCo',
-          path: 'pages/dashboard/index',
-          extraData: {},
-        })
-      },
-    },
-  ]
-  return (
-    <FooterWrap>
-      {list.map((item) => (
-        <Item key={item.title} {...item} />
-      ))}
-    </FooterWrap>
-  )
-}
+
 export default function Index() {
   return (
-    <Providers>
-      <ErrorBoundary name="swap">
-        <Swap />
-        <FooterMenu />
-      </ErrorBoundary>
-    </Providers>
+    <ErrorBoundary name="swap">
+      <Swap />
+    </ErrorBoundary>
   )
 }
